+++ crates/pathfinder/src/bin/pathfinder/config/cli.rs
@@ -6,7 +6,6 @@ use crate::config::builder::ConfigBuilder;
 
 use super::ConfigOption;
 
-const CONFIG_KEY: &str = "config";
 const DATA_DIR_KEY: &str = "data-directory";
 const ETH_URL_KEY: &str = "ethereum.url";
 const ETH_PASS_KEY: &str = "ethereum.password";
@@ -23,12 +22,11 @@ const GATEWAY: &str = "gateway-url";
 const FEEDER_GATEWAY: &str = "feeder-gateway-url";
 const CHAIN_ID: &str = "chain-id";
 
-/// Parses the cmd line arguments and returns the optional
-/// configuration file's path and the specified configuration options.
+/// Parses the cmd line arguments and returns the specified configuration options.
 ///
 /// Note: This will terminate the program if invalid arguments are supplied.
 ///       This is intended, as [clap] will show the program usage / help.
-pub fn parse_cmd_line() -> (Option<String>, ConfigBuilder) {
+pub fn parse_cmd_line() -> ConfigBuilder {
     // A thin wrapper around `parse_args()`. This should be kept thin
     // to enable test coverage without requiring cmd line arg input.
     match parse_args(std::env::args_os()) {
@@ -39,14 +37,13 @@ pub fn parse_cmd_line() -> (Option<String>, ConfigBuilder) {
 
 /// A wrapper around [clap::Command]'s `get_matches_from_safe()` which returns
 /// a [ConfigOption].
-fn parse_args<I, T>(args: I) -> clap::Result<(Option<String>, ConfigBuilder)>
+fn parse_args<I, T>(args: I) -> clap::Result<ConfigBuilder>
 where
     I: IntoIterator<Item = T>,
     T: Into<OsString> + Clone,
 {
     let args = clap_app().try_get_matches_from(args)?;
 
-    let config_filepath = args.value_of(CONFIG_KEY).map(|s| s.to_owned());
     let data_directory = args.value_of(DATA_DIR_KEY).map(|s| s.to_owned());
     let ethereum_url = args.value_of(ETH_URL_KEY).map(|s| s.to_owned());
     let ethereum_password = args.value_of(ETH_PASS_KEY).map(|s| s.to_owned());
@@ -81,7 +78,7 @@ where
         .with(ConfigOption::FeederGatewayUrl, feeder_gateway)
         .with(ConfigOption::ChainId, chain_id);
 
-    Ok((config_filepath, cfg))
+    Ok(cfg)
 }
 
 /// Defines our command-line interface using [clap::Command].
@@ -98,14 +95,6 @@ fn clap_app() -> clap::Command<'static> {
     clap::Command::new("Pathfinder")
         .version(version)
         .about("A StarkNet node implemented by Equilibrium. Submit bug reports and issues at https://github.com/eqlabs/pathfinder.")
-        .arg(
-            Arg::new(CONFIG_KEY)
-                .short('c')
-                .long(CONFIG_KEY)
-                .help("Path to the configuration file.")
-                .value_name("FILE")
-                .takes_value(true),
-        )
         .arg(
             Arg::new(ETH_PASS_KEY)
                 .long(ETH_PASS_KEY)
@@ -274,7 +263,7 @@ mod tests {
         clear_environment();
 
         let value = "value".to_owned();
-        let (_, mut cfg) = parse_args(vec!["bin name", "--ethereum.url", &value]).unwrap();
+        let mut cfg = parse_args(vec!["bin name", "--ethereum.url", &value]).unwrap();
         assert_eq!(cfg.take(ConfigOption::EthereumHttpUrl), Some(value));
     }
 
@@ -285,7 +274,7 @@ mod tests {
 
         let value = "value".to_owned();
         env::set_var("PATHFINDER_ETHEREUM_API_URL", &value);
-        let (_, mut cfg) = parse_args(vec!["bin name"]).unwrap();
+        let mut cfg = parse_args(vec!["bin name"]).unwrap();
         assert_eq!(cfg.take(ConfigOption::EthereumHttpUrl), Some(value));
     }
 
@@ -295,7 +284,7 @@ mod tests {
         clear_environment();
 
         let value = "value".to_owned();
-        let (_, mut cfg) = parse_args(vec!["bin name", "--ethereum.password", &value]).unwrap();
+        let mut cfg = parse_args(vec!["bin name", "--ethereum.password", &value]).unwrap();
         assert_eq!(cfg.take(ConfigOption::EthereumPassword), Some(value));
     }
 
@@ -306,28 +295,8 @@ mod tests {
 
         let value = "value".to_owned();
         env::set_var("PATHFINDER_ETHEREUM_API_PASSWORD", &value);
-        let (_, mut cfg) = parse_args(vec!["bin name"]).unwrap();
+        let mut cfg = parse_args(vec!["bin name"]).unwrap();
         assert_eq!(cfg.take(ConfigOption::EthereumPassword), Some(value));
-    }
-
-    #[test]
-    fn config_filepath_short() {
-        let _env_guard = ENV_VAR_MUTEX.lock().unwrap_or_else(|e| e.into_inner());
-        clear_environment();
-
-        let value = "value".to_owned();
-        let (filepath, _) = parse_args(vec!["bin name", "-c", &value]).unwrap();
-        assert_eq!(filepath, Some(value));
-    }
-
-    #[test]
-    fn config_filepath_long() {
-        let _env_guard = ENV_VAR_MUTEX.lock().unwrap_or_else(|e| e.into_inner());
-        clear_environment();
-
-        let value = "value".to_owned();
-        let (filepath, _) = parse_args(vec!["bin name", "--config", &value]).unwrap();
-        assert_eq!(filepath, Some(value));
     }
 
     #[test]
@@ -336,7 +305,7 @@ mod tests {
         clear_environment();
 
         let value = "value".to_owned();
-        let (_, mut cfg) = parse_args(vec!["bin name", "--http-rpc", &value]).unwrap();
+        let mut cfg = parse_args(vec!["bin name", "--http-rpc", &value]).unwrap();
         assert_eq!(cfg.take(ConfigOption::HttpRpcAddress), Some(value));
     }
 
@@ -347,7 +316,7 @@ mod tests {
 
         let value = "value".to_owned();
         env::set_var("PATHFINDER_HTTP_RPC_ADDRESS", &value);
-        let (_, mut cfg) = parse_args(vec!["bin name"]).unwrap();
+        let mut cfg = parse_args(vec!["bin name"]).unwrap();
         assert_eq!(cfg.take(ConfigOption::HttpRpcAddress), Some(value));
     }
 
@@ -357,7 +326,7 @@ mod tests {
         clear_environment();
 
         let value = "value".to_owned();
-        let (_, mut cfg) = parse_args(vec!["bin name", "--data-directory", &value]).unwrap();
+        let mut cfg = parse_args(vec!["bin name", "--data-directory", &value]).unwrap();
         assert_eq!(cfg.take(ConfigOption::DataDirectory), Some(value));
     }
 
@@ -368,7 +337,7 @@ mod tests {
 
         let value = "value".to_owned();
         env::set_var("PATHFINDER_DATA_DIRECTORY", &value);
-        let (_, mut cfg) = parse_args(vec!["bin name"]).unwrap();
+        let mut cfg = parse_args(vec!["bin name"]).unwrap();
         assert_eq!(cfg.take(ConfigOption::DataDirectory), Some(value));
     }
 
@@ -378,7 +347,7 @@ mod tests {
         clear_environment();
 
         let value = "value".to_owned();
-        let (_, mut cfg) = parse_args(vec!["bin name", "--sequencer-url", &value]).unwrap();
+        let mut cfg = parse_args(vec!["bin name", "--sequencer-url", &value]).unwrap();
         assert_eq!(cfg.take(ConfigOption::SequencerHttpUrl), Some(value));
     }
 
@@ -389,7 +358,7 @@ mod tests {
 
         let value = "value".to_owned();
         env::set_var("PATHFINDER_SEQUENCER_URL", &value);
-        let (_, mut cfg) = parse_args(vec!["bin name"]).unwrap();
+        let mut cfg = parse_args(vec!["bin name"]).unwrap();
         assert_eq!(cfg.take(ConfigOption::SequencerHttpUrl), Some(value));
     }
 
@@ -399,7 +368,7 @@ mod tests {
         clear_environment();
 
         let value = "value".to_owned();
-        let (_, mut cfg) = parse_args(vec!["bin name", "--python-subprocesses", &value]).unwrap();
+        let mut cfg = parse_args(vec!["bin name", "--python-subprocesses", &value]).unwrap();
         assert_eq!(cfg.take(ConfigOption::PythonSubprocesses), Some(value));
     }
 
@@ -410,7 +379,7 @@ mod tests {
 
         let value = "value".to_owned();
         env::set_var("PATHFINDER_PYTHON_SUBPROCESSES", &value);
-        let (_, mut cfg) = parse_args(vec!["bin name"]).unwrap();
+        let mut cfg = parse_args(vec!["bin name"]).unwrap();
         assert_eq!(cfg.take(ConfigOption::PythonSubprocesses), Some(value));
     }
 
@@ -420,7 +389,7 @@ mod tests {
         clear_environment();
 
         let value = "value".to_owned();
-        let (_, mut cfg) = parse_args(vec!["bin name", "--sqlite-wal", &value]).unwrap();
+        let mut cfg = parse_args(vec!["bin name", "--sqlite-wal", &value]).unwrap();
         assert_eq!(
             cfg.take(ConfigOption::EnableSQLiteWriteAheadLogging),
             Some(value)
@@ -434,7 +403,7 @@ mod tests {
 
         let value = "value".to_owned();
         env::set_var("PATHFINDER_SQLITE_WAL", &value);
-        let (_, mut cfg) = parse_args(vec!["bin name"]).unwrap();
+        let mut cfg = parse_args(vec!["bin name"]).unwrap();
         assert_eq!(
             cfg.take(ConfigOption::EnableSQLiteWriteAheadLogging),
             Some(value)
@@ -447,7 +416,7 @@ mod tests {
         clear_environment();
 
         let value = "value".to_owned();
-        let (_, mut cfg) = parse_args(vec!["bin name", "--poll-pending", &value]).unwrap();
+        let mut cfg = parse_args(vec!["bin name", "--poll-pending", &value]).unwrap();
         assert_eq!(cfg.take(ConfigOption::PollPending), Some(value));
     }
 
@@ -458,7 +427,7 @@ mod tests {
 
         let value = "value".to_owned();
         env::set_var("PATHFINDER_POLL_PENDING", &value);
-        let (_, mut cfg) = parse_args(vec!["bin name"]).unwrap();
+        let mut cfg = parse_args(vec!["bin name"]).unwrap();
         assert_eq!(cfg.take(ConfigOption::PollPending), Some(value));
     }
 
@@ -468,7 +437,7 @@ mod tests {
         clear_environment();
 
         let value = "value".to_owned();
-        let (_, mut cfg) = parse_args(vec!["bin name", "--monitor-address", &value]).unwrap();
+        let mut cfg = parse_args(vec!["bin name", "--monitor-address", &value]).unwrap();
         assert_eq!(cfg.take(ConfigOption::MonitorAddress), Some(value));
     }
 
@@ -479,7 +448,7 @@ mod tests {
 
         let value = "value".to_owned();
         env::set_var("PATHFINDER_MONITOR_ADDRESS", &value);
-        let (_, mut cfg) = parse_args(vec!["bin name"]).unwrap();
+        let mut cfg = parse_args(vec!["bin name"]).unwrap();
         assert_eq!(cfg.take(ConfigOption::MonitorAddress), Some(value));
     }
 
@@ -488,7 +457,7 @@ mod tests {
         let _env_guard = ENV_VAR_MUTEX.lock().unwrap_or_else(|e| e.into_inner());
         clear_environment();
 
-        let (_, mut cfg) = parse_args(vec!["bin name", "--testnet2"]).unwrap();
+        let mut cfg = parse_args(vec!["bin name", "--testnet2"]).unwrap();
         assert_eq!(cfg.take(ConfigOption::Testnet2), Some("".to_owned()));
     }
 
@@ -498,12 +467,12 @@ mod tests {
         clear_environment();
 
         let value = "mainnet".to_owned();
-        let (_, mut cfg) = parse_args(vec!["bin name", "--network", &value]).unwrap();
+        let mut cfg = parse_args(vec!["bin name", "--network", &value]).unwrap();
         assert_eq!(cfg.take(ConfigOption::Network), Some(value));
 
         let value = "mainnet".to_owned();
         env::set_var("PATHFINDER_NETWORK", &value);
-        let (_, mut cfg) = parse_args(vec!["bin name"]).unwrap();
+        let mut cfg = parse_args(vec!["bin name"]).unwrap();
         assert_eq!(cfg.take(ConfigOption::Network), Some(value));
     }
 
@@ -513,12 +482,12 @@ mod tests {
         clear_environment();
 
         let value = "value".to_owned();
-        let (_, mut cfg) = parse_args(vec!["bin name", "--gateway-url", &value]).unwrap();
+        let mut cfg = parse_args(vec!["bin name", "--gateway-url", &value]).unwrap();
         assert_eq!(cfg.take(ConfigOption::GatewayUrl), Some(value));
 
         let value = "value".to_owned();
         env::set_var("PATHFINDER_GATEWAY_URL", &value);
-        let (_, mut cfg) = parse_args(vec!["bin name"]).unwrap();
+        let mut cfg = parse_args(vec!["bin name"]).unwrap();
         assert_eq!(cfg.take(ConfigOption::GatewayUrl), Some(value));
     }
 
@@ -528,12 +497,12 @@ mod tests {
         clear_environment();
 
         let value = "value".to_owned();
-        let (_, mut cfg) = parse_args(vec!["bin name", "--feeder-gateway-url", &value]).unwrap();
+        let mut cfg = parse_args(vec!["bin name", "--feeder-gateway-url", &value]).unwrap();
         assert_eq!(cfg.take(ConfigOption::FeederGatewayUrl), Some(value));
 
         let value = "value".to_owned();
         env::set_var("PATHFINDER_FEEDER_GATEWAY_URL", &value);
-        let (_, mut cfg) = parse_args(vec!["bin name"]).unwrap();
+        let mut cfg = parse_args(vec!["bin name"]).unwrap();
         assert_eq!(cfg.take(ConfigOption::FeederGatewayUrl), Some(value));
     }
 
@@ -543,12 +512,12 @@ mod tests {
         clear_environment();
 
         let value = "value".to_owned();
-        let (_, mut cfg) = parse_args(vec!["bin name", "--chain-id", &value]).unwrap();
+        let mut cfg = parse_args(vec!["bin name", "--chain-id", &value]).unwrap();
         assert_eq!(cfg.take(ConfigOption::ChainId), Some(value));
 
         let value = "value".to_owned();
         env::set_var("PATHFINDER_CHAIN_ID", &value);
-        let (_, mut cfg) = parse_args(vec!["bin name"]).unwrap();
+        let mut cfg = parse_args(vec!["bin name"]).unwrap();
         assert_eq!(cfg.take(ConfigOption::ChainId), Some(value));
     }
 
@@ -557,8 +526,7 @@ mod tests {
         let _env_guard = ENV_VAR_MUTEX.lock().unwrap_or_else(|e| e.into_inner());
         clear_environment();
 
-        let (filepath, cfg) = parse_args(vec!["bin name"]).unwrap();
-        assert_eq!(filepath, None);
+        let cfg = parse_args(vec!["bin name"]).unwrap();
         assert_eq!(cfg, ConfigBuilder::default());
     }
 }